refactor: move v2 related to the v2 folder
